rag-chatbot/requirements.txt
langchain>=0.2
langchain-community>=0.2
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.0.1
faiss-cpu
huggingface_hub
transformers
sentence-transformers
pypdf
python-dotenv